requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==45.0.7
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.8.0
django-filter==25.1
django-otp==1.6.1
django-recaptcha==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.10.0
